pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
description = "NSE algorithmic trading system — Zerodha Kite + XGBoost + FinBERT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core
    "pandas>=2.2.0",
    "numpy>=1.26.4",
    "scipy>=1.12.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "apscheduler>=3.10.4",
    "structlog>=24.1.0",

    # Market data
    "kiteconnect>=5.0.1",
    "finnhub-python>=2.4.19",
    "feedparser>=6.0.11",

    # Technical analysis
    "pandas-ta>=0.3.14b",

    # ML
    "lightgbm>=4.3.0",
    "xgboost>=2.0.3",
    "scikit-learn>=1.4.1",
    "hmmlearn>=0.3.2",
    "optuna>=3.6.1",
    "shap>=0.45.0",

    # NLP / LLM
    "transformers>=4.39.0",
    "torch>=2.2.1",
    "sentence-transformers>=2.5.1",

    # Database
    "alembic>=1.13.0",
    "sqlalchemy>=2.0.28",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.3",

    # Monitoring
    "mlflow>=2.11.1",
    "prometheus-client>=0.20.0",
    "python-telegram-bot>=21.1",

    # Backtesting
    "vectorbt>=0.26.2",
]

[dependency-groups]
dev = [
    # Testing
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    # Linting / formatting
    "ruff>=0.9.0",
    # Type checking
    "mypy>=1.14.0",
    "types-redis>=4.6.0",
    "types-PyYAML>=6.0.12",
    # Security
    "bandit[toml]>=1.8.0",
    "pip-audit>=2.8.0",
    # Git hooks
    "pre-commit>=4.1.0",
]

[project.optional-dependencies]
upstox = [
    # requests is already a transitive dep via kiteconnect; listed explicitly
    # so uv knows it's required for the upstox extra.
    # websocket-client is also already present via kiteconnect.
    # For live-streaming protobuf decoding, install the SDK from GitHub:
    #   pip install git+https://github.com/upstox/upstox-python.git
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = [
    "main.py",
    "migrations/*",
]

[tool.hatch.build.targets.wheel]
packages = ["data", "signals", "llm", "risk", "execution", "backtest", "options", "orchestrator", "monitoring"]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes (undefined names, unused imports)
    "I",    # isort (import order)
    "UP",   # pyupgrade (modernise syntax)
    "B",    # flake8-bugbear (common bugs)
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "PIE",  # flake8-pie (misc improvements)
    "RUF",  # Ruff-native rules
]
ignore = [
    "E501",   # line-too-long — handled by the formatter
    "UP007",  # Use X | Y — keep Optional[] for Pydantic v1 compat
    "UP035",  # typing module deprecated — some third-party stubs still need it
    "SIM108", # ternary operator — can hurt readability in complex conditions
    "B008",   # function-call-in-default-argument — valid with Pydantic validators
    "RUF001", # string contains ambiguous unicode — intentional in financial literals (×, ₹)
    "RUF002", # docstring contains ambiguous unicode — intentional in math/finance docs (×, –)
    "RUF003", # comment contains ambiguous unicode — intentional in math/finance comments
    "RUF012", # mutable class vars — handled by dataclass/pydantic
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "B011",   # assert False can be used as TODO marker
    "B017",   # pytest.raises with no match arg
    "S101",   # assert in tests is fine
    "F841",   # unused local vars common in test setup (mock captures)
    "RUF059", # unpacked variables not used (mock return value capturing)
    "B905",   # zip strict= not required in tests
]
"backtest/**" = [
    "S311",   # random in simulations is intentional
    "E402",   # module-level imports after sys.path manipulation
]
"notebooks/**" = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ---------------------------------------------------------------------------
# Mypy — type checking
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_configs = true
warn_return_any = false        # too noisy with untyped third-party libs
disallow_untyped_defs = false  # incrementally adopted — enable per-module later
exclude = [
    "migrations/",
    "notebooks/",
    "data/notebooks/",
]

# ---------------------------------------------------------------------------
# Bandit — security linting
# ---------------------------------------------------------------------------

[tool.bandit]
exclude_dirs = ["tests", "notebooks", "migrations", ".venv"]
skips = [
    "B101",  # assert_used — fine in test helpers
    "B110",  # try_except_pass — intentional in monitoring/alerting code for resilience
    "B311",  # random — intentional in backtests / simulations
    "B603",  # subprocess_without_shell_equals_true — not used here
    "B607",  # start_process_with_partial_path — not used here
]
